Add comments
Added a couple of comments explaining our changes. We created a "dumbbell" object that, when lifted, increases the player's strength. If a player is strong, they cannot be cursed, instead their strength will be reduced. We also created a "chocolate" object. If the player has a chocolate, the witch will take their chocolate instead of turning them into a frog.